reverted to cuda

+++ data_loader.py
@@ -38,10 +38,10 @@ class TrainUtils:
 
     def trainTestSplit(self, sig, label, trainPercent, shuffle=False):
         X_train, X_test, y_train, y_test = train_test_split(sig, label, train_size=trainPercent, shuffle=shuffle)
-        X_train = np.array(X_train, dtype=np.float32)
-        X_test = np.array(X_test, dtype=np.float32)
-        y_train = np.array(y_train, dtype=np.float32)
-        y_test = np.array(y_test, dtype=np.float32)
+        X_train = np.array(X_train)
+        X_test = np.array(X_test)
+        y_train = np.array(y_train)
+        y_test = np.array(y_test)
         return X_train, X_test, y_train, y_test
 
 
@@ -56,10 +56,10 @@ class Data_Loader():
         ecgWindows, fecgWindows,fqrs_rpeaks = self.trainUtils.prepareData(delay=5)
         X_train, X_test, Y_train, Y_test = self.trainUtils.trainTestSplit(ecgWindows, fecgWindows, len(ecgWindows)-1)
 
-        X_train = np.reshape(X_train, [-1, X_train.shape[2], X_train.shape[1]]).astype(np.float32, copy=False)
-        X_test = np.reshape(X_test, [-1, X_test.shape[2], X_test.shape[1]]).astype(np.float32, copy=False)
-        Y_train = np.reshape(Y_train, [-1, Y_train.shape[2], Y_train.shape[1]]).astype(np.float32, copy=False)
-        Y_test = np.reshape(Y_test, [-1, Y_test.shape[2], Y_test.shape[1]]).astype(np.float32, copy=False)
+        X_train = np.reshape(X_train, [-1, X_train.shape[2], X_train.shape[1]])
+        X_test = np.reshape(X_test, [-1, X_test.shape[2], X_test.shape[1]])
+        Y_train = np.reshape(Y_train, [-1, Y_train.shape[2], Y_train.shape[1]])
+        Y_test = np.reshape(Y_test, [-1, Y_test.shape[2], Y_test.shape[1]])
         # print(X_train.shape)
         
         
@@ -149,11 +149,11 @@ class FECGDataset(Dataset):
             
             
             min_max_scaler = MinMaxScaler(feature_range=(-1, 1), copy=False)           
-            MECG_signal = min_max_scaler.fit_transform(MECG_signal.transpose()).transpose().astype(np.float32, copy=False)
+            MECG_signal = min_max_scaler.fit_transform(MECG_signal.transpose()).transpose()
 
-            AECG_signal = min_max_scaler.fit_transform(self.X_train[index,:,:].transpose()).transpose().astype(np.float32, copy=False)
-            FECG_signal = min_max_scaler.fit_transform(self.Y_train[index,:,:].transpose()).transpose().astype(np.float32, copy=False)
-            BIAS_signal = min_max_scaler.fit_transform(noise.transpose()).transpose().astype(np.float32, copy=False)
+            AECG_signal = min_max_scaler.fit_transform(self.X_train[index,:,:].transpose()).transpose()
+            FECG_signal = min_max_scaler.fit_transform(self.Y_train[index,:,:].transpose()).transpose()
+            BIAS_signal = min_max_scaler.fit_transform(noise.transpose()).transpose()
             
             
             return AECG_signal,FECG_signal,MECG_signal,BIAS_signal
@@ -210,12 +210,7 @@ class FECGDataset(Dataset):
             noise = MECG_signal.copy()
             noise[0,int(max(0,M_index-15)):int(min(MECG_signal.shape[-1],M_index+15))] = noise[0,max(0,M_index-15)]
             
-            return (
-                self.X_test[index,:,:].astype(np.float32, copy=False),
-                self.Y_test[index,:,:].astype(np.float32, copy=False),
-                MECG_signal.astype(np.float32, copy=False),
-                noise.astype(np.float32, copy=False),
-            )
+            return self.X_test[index,:,:],self.Y_test[index,:,:],MECG_signal,noise
         
     def __len__(self):
         return len(self.X_train)
